backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
openai-whisper>=20231117
ffmpeg-python>=0.2.0
python-multipart>=0.0.9
